Adapt old tests and blank the tests that are not working
